refactor/auth: 회원가입 Auth 도메인으로 옮기기
