+++ src/main/java/org/example/queristia/domain/auth/dto/UserSaveResponse.java
@@ -1,4 +1,4 @@
-package org.example.queristia.domain.user.dto;
+package org.example.queristia.domain.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
