+++ src/main/java/org/example/queristia/domain/user/controller/UserController.java
@@ -2,12 +2,7 @@ package org.example.queristia.domain.user.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.queristia.domain.user.dto.UserSaveRequest;
-import org.example.queristia.domain.user.dto.UserSaveResponse;
 import org.example.queristia.domain.user.service.UserService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -20,9 +15,5 @@ public class UserController {
 
     private final UserService userService;
 
-    @PostMapping("/signup")
-    public ResponseEntity<UserSaveResponse> Signup(@RequestBody UserSaveRequest userSaveRequest) {
 
-        return ResponseEntity.ok().body(userService.signup(userSaveRequest));
-    }
 }
